bfam_qa/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
python-dotenv==1.0.0
google-generativeai==0.3.1
markdown2==2.4.10
openai>=1.0.0

